+++ .homeycompose/app.json
@@ -2,8 +2,7 @@
 	"id": "com.netscan",
 	"version": "2.0.9",
 	"name": {
-		"en": "Net Scan",
-		"nl": "Net Scan"
+		"en": "Net Scan"
 	},
 	"platforms": [
 		"local"
@@ -19,7 +18,8 @@
 	"description": {
 		"en": "Watch devices on your local network by scanning for TCP ports",
 		"nl": "Houdt apparaten op uw netwerk in de gaten middels een TCP poort",
-		"de": "Überwachen Sie Geräte in Ihrem lokalen Netzwerk durch Scannen der TCP-Ports"
+		"de": "Überwachen Sie Geräte in Ihrem lokalen Netzwerk durch Scannen der TCP-Ports",
+		"pl": "Obserwuj urządzenia w lokalnej sieci poprzez skanowanie portów TCP"
 	},
 	"contributors": {
 		"developers": [
