+++ lib/device.js
@@ -2,7 +2,9 @@
 'use strict';
 
 const Homey = require('homey');
+const utils = require('./utils')
 var net = require("net");
+
 
 class TcpIpDevice extends Homey.Device {
     async onInit() {
@@ -110,7 +112,7 @@ class TcpIpDevice extends Homey.Device {
             socket.once('error', (error) => {
                 const reason = error?.code
                     ? (map[error.code] || `unknown code: ${error.code}`)
-                    : `unknown error: ${this.homey.app.varToString(error)}`;
+                    : `unknown error: ${utils.varToString(error)}`;
                 // EALREADY => treat as online
                 done(reason === 'ready', reason);
             });
